Plant-Disease-Detection-DL/requirements.txt
# TensorFlow and Deep Learning
tensorflow>=2.12.0
tensorflow-hub>=0.13.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Application
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Jupyter Notebook (optional for development)
jupyter>=1.0.0
ipykernel>=6.20.0
ipywidgets>=8.0.0

# Additional Tools
plotly>=5.14.0
